% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_conversion.R
\name{BATCH_prep}
\alias{BATCH_prep}
\title{BATCH_prep}
\usage{
BATCH_prep(directory, exclude = c("FIRST README.txt", "Gassampler_log.txt"))
}
\arguments{
\item{directory}{folder containing the raw SEMACHaqua file}

\item{overwrite}{Should raw file be overwritten with converted structure, or should
converted file be saved as .../prepped_filename}
}
\description{
Converter for "new" SEMACH aqua data format in SEMACHcalc readable format - apllies conversion to folder with data
}
