% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_processor.R
\name{batch_processor}
\alias{batch_processor}
\title{Batch processor}
\usage{
batch_processor(
  directory,
  filenames,
  method = "default",
  start = NA,
  end = NA,
  A_chamber = NA,
  V_chamber = NA,
  metafile,
  draw = c("none"),
  save_plot = F,
  show_plot = F,
  draw.method = "svg",
  tz = "America/Manaus"
)
}
\arguments{
\item{directory}{Path of folder containing data to be processed.}

\item{filenames}{Vector of actual filenames of the .txt-files to be processed. Only filename, do NOT add .txt (e.g., "filename.txt")}

\item{method}{Either "default" (also literally default), "fixed", or "individual". Default uses default chamber proportions (area, volume) and time intervals (1min-5min).
Fixed allows customisation of several or all parameters to a fixed value different to the default. Input through A_chamber, V_chamber, start and end. If no parameters are edited, use default instead.
Individual allows for individual parameter customisation for each .txt-file. Requires input of metadata file. If no parameters are edited, use default instead.}

\item{start}{Starting point of evaluation in seconds after start of recording. Used for method="fixed"}

\item{end}{End point of evaluation in seconds after start of recording. Used for method="fixed"}

\item{A_chamber}{Chamber base area in m². Implemented for completion, usually not changed. Used for method="fixed"}

\item{V_chamber}{Chamber volume in m³. Useful e.g., when chamber volume is extended by base ring. Used for method="fixed"}

\item{metafile}{data frame containing 5 columns:
1 - 'filenames' As used in filenames parameter. Each measurement file has it's own parameter row.
2 - 'V_chamber' Deviating chamber volume in m³ for each measurement file. If default, set to NA or enter default manually.
3 - 'A_chamber' Deviating chamber area in m² for each measurement file. If default, set to NA or enter default manually.
4 - 'start' Starting point of evaluation in seconds after start of recording for each measurement file. If default, set to NA or enter default manually.
5 - 'end' Endpoint of evaluation in seconds after start of recording for each measurement file. If default, set to NA or enter default manually.}

\item{draw}{Which variables should be included in plot? Default is c("none").
Other possible inputs are vector containing variable names
("rH_i","T_i","P_i","SM","ST","CO2","rH_o","T_o","P_o","PAR"). To plot all, use c("all") instead.}

\item{save_plot}{Should the plot be saved? T/F Default is F.}

\item{show_plot}{Should the plot be drawn on console window T/F Default is F.}

\item{draw.method}{Which format should the plots be drawn in? Default is svg, other option is jpeg (lower res.).
Using c("svg","jpeg") allows to draw both formats at once.
#' @param tz Timezone of timestamps}
}
\description{
Allows quick processing of multiple SEMACH-output-files. Either default parameters (1min-5min)
or custom time interval and chamber dimension settings. All .txt-files have to be situated in the same folder. Returns dataframe with averages over selected measurement timeframe, CO2 accumulation rate (ppm/min) and CO2 flux (µmol m-2 s-1)
}
