% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_plot.R
\name{data_plot}
\alias{data_plot}
\title{data_plot}
\usage{
data_plot(
  xx,
  filepath,
  draw = c("all"),
  save_plot = F,
  show_plot = T,
  draw.method = "svg"
)
}
\arguments{
\item{xx}{Subset of SEMACH raw data to be drawn as internally created by flux.calc. If the function is used
manually, pease be sure to add t_sec column, with the respective measurement time in seconds. Furthermore the columns
as described for draw should be included and named accordingly.}

\item{filepath}{Path from which data is loaded. E.g., "E:/Folder/measurement.txt"}

\item{draw}{Which variables should be included in plot? Default is c("all").
Other possible inputs are vector containing variable names.
("CO2","rH_i","T_i","P_i","rH_o","T_o","P_o","PAR","WT"), or empty vector c().}

\item{save_plot}{Should the plot be saved? T/F Default is T.}

\item{show_plot}{Should the plot be drawn on console window T/F Default is F.}

\item{draw.method}{Which format should the plots be drawn in? Default is svg, other option is jpeg (lower res.).
Using c("svg","jpeg") allows to draw both formats at once.}
}
\description{
Plots raw SEMACH data and saves it (as svg) and / or displays it on console if desired.
Mostly for internal use.
}
