% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate}
\usage{
evaluate(
  directory,
  save = TRUE,
  draw = c("none"),
  save_plot = F,
  show_plot = F,
  draw.method = "svg",
  tz = "America/Manaus"
)
}
\arguments{
\item{directory}{Folder directory of folder that contains data and metafile. Example: "C:/Project/Data/Folder" or "C:/Project/Data/Folder/".}

\item{save}{Should summary.csv be created in directory? Default == TRUE.}

\item{draw}{Which variables should be included in plot? Default is c("none").
Other possible inputs are vector containing variable names
("rH_i","T_i","P_i","SM","ST","CO2","rH_o","T_o","P_o","PAR"). To plot all, use c("all") instead.}

\item{save_plot}{Should the plot be saved? T/F Default is F.}

\item{show_plot}{Should the plot be drawn on console window T/F Default is F.}

\item{draw.method}{Which format should the plots be drawn in? Default is svg, other option is jpeg (lower res.).
Using c("svg","jpeg") allows to draw both formats at once.}

\item{tz}{Timezone of timestamps}
}
\description{
Takes folder directory containing SEMACH raw data files (.txt) AND a metadata file called "metadata" (.xlsx).
The metadata file should be constructed as described in SMEACHcalc::batch_processor and contain all raw data files in the
folder (which the user wants to evaluate) without the .txt ending. The sheet name in the excel file has to be "Tabelle1". The function calculates flux and averages as described in SMACHcalc::batch_processor
for each raw file and returns values as data frame. When save == TRUE also creates csv-file called "summary" in the directory.
}
