% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux_calc_aqua.R
\name{flux_calc_aqua}
\alias{flux_calc_aqua}
\title{Flux calculator}
\usage{
flux_calc_aqua(
  filepath,
  start = 61,
  end = 300,
  V_chamber = 0.01587,
  A_chamber = 0.0491,
  draw = c("all"),
  save_plot = F,
  show_plot = T,
  draw.method = "svg",
  tz = "America/Manaus"
)
}
\arguments{
\item{filepath}{Path from which data is loaded. E.g., "E:/Folder/measurement.csv"}

\item{start}{Starting point of evaluation in seconds after start of recording. Default is 61 s (or 1 min)}

\item{end}{End point of evaluation in seconds after start of recording. Default is 300 s (or 5 min)}

\item{V_chamber}{Chamber volume in m³. Useful e.g., when chamber volume is extended by base ring.
!!! Default is still land chamber dimensions 0.01587 m³ !!!}

\item{A_chamber}{Chamber base area in m². Implemented for completion, usually not changed.
!!!  Default is still land chamber dimensions 0.0491 m² !!!}

\item{draw}{Which variables should be included in plot? Default is c("all").
Other possible inputs are vector containing variable names or c("none") if no plotting is desired.
("CO2","rH_i","T_i","P_i","rH_o","T_o","P_o","PAR","WT"), or empty vector c().}

\item{save_plot}{Should the plot be saved? T/F Default is T.}

\item{show_plot}{Should the plot be drawn on console window T/F Default is F.}

\item{draw.method}{Which format should the plots be drawn in? Default is svg, other option is jpeg (lower res.).
Using c("svg","jpeg") allows to draw both formats at once.}

\item{tz}{Timezone of timestamps}
}
\description{
Calculates CO2 flux from SEMACHaqua raw data (.csv format). Returns vector with averages, dCO2 in ppm/min and flux in µmol m-2 s-1.
}
