% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_conversion.R
\name{SEMACH_aqua_data_prep}
\alias{SEMACH_aqua_data_prep}
\title{SEMACH_aqua_data_prep}
\usage{
SEMACH_aqua_data_prep(directory, overwrite = F)
}
\arguments{
\item{directory}{filepath of the raw SEMACHaqua file}

\item{overwrite}{Should raw file be overwritten with converted structure, or should
converted file be saved as .../prepped_filename}
}
\description{
Converter for "new" SEMACH aqua data format in SEMACHcalc readable format
}
